requirements.txt
streamlit>=1.32,<2.0
pillow>=10.2
opencv-python-headless>=4.9
numpy>=1.26,<2.0
onnxruntime>=1.17
rembg>=2.0.56
openai>=1.0
